Added "anonymous" user to the database.

Conflicts:
	src/main/java/org/studentsuccessplan/mygps/business/SelfHelpGuideManager.java
	src/main/java/org/studentsuccessplan/mygps/business/TaskManager.java
	src/main/java/org/studentsuccessplan/mygps/web/MyGpsChallengeController.java
	src/main/java/org/studentsuccessplan/ssp/transferobject/reference/ChallengeTO.java

+++ src/test/java/org/studentsuccessplan/ssp/dao/reference/ChallengeReferralDaoTest.java
@@ -17,7 +17,6 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.transaction.TransactionConfiguration;
 import org.springframework.transaction.annotation.Transactional;
-
 import org.studentsuccessplan.ssp.dao.PersonDao;
 import org.studentsuccessplan.ssp.model.ObjectStatus;
 import org.studentsuccessplan.ssp.model.Person;
@@ -135,7 +134,7 @@ public class ChallengeReferralDaoTest {
 
 	@Test
 	public void countByChallengeIdNotOnActiveTaskList() {
-		long count =
+		int count =
 				dao.countByChallengeIdNotOnActiveTaskList(
 						testChallenge.getId(), testStudent, "testSessionId");
 		assertTrue(count > -1);
